+++ spec/actarray_spec.rb
@@ -1,7 +1,7 @@
 require File.dirname(__FILE__) + "/spec_helper"
 
 include Player
-describe Player::Power do
+describe Player::ActArray do
   before do
     client = mock_client
 
@@ -15,9 +15,10 @@ describe Player::Power do
 
   it 'should set power state for all actuators' do
     should_send_message(PLAYER_MSGTYPE_REQ, PLAYER_ACTARRAY_REQ_POWER, [0].pack("N"))
-    @actarray.turn_off!
+    @actarray.power_off!
 
     should_send_message(PLAYER_MSGTYPE_REQ, PLAYER_ACTARRAY_REQ_POWER, [1].pack("N"))
-    @actarray.turn_on!
+    @actarray.power_on!
   end
+
 end